+++ src/helpers/constants.ts
@@ -4,4 +4,19 @@ export namespace Constants {
     export const COMMAND_DISABLE_UPLOAD = 'extension.prophet.command.disable.upload';
     export const COMMAND_ENABLE_UPLOAD = 'extension.prophet.command.enable.upload';
     export const TERMINAL_NAME = 'Build Prophet';
+    export const QUICKPICK_TITLE_HOSTNAME = 'Select the Hostname';
+    export const QUICKPICK_TITLE_CODEVERSON = 'Select the Code Version';
+    export const HOSTNAME_HISTORY_PROPERTY = 'sfcc-dw-helper.hostnameHistory';
+    export const CODEVERSION_HISTORY_PROPERTY = 'sfcc-dw-helper.codeversionHistory';
+    export const HOSTNAME = 'hostname';
+    export const CODEVERSION = 'code-version';
+    export const HOSTNAME_HISTORY_PROPERTY_SHORT = 'hostnameHistory';
+    export const CODEVERSION_HISTORY_PROPERTY_SHORT = 'codeversionHistory';
+    export const HOSTNAME_INFO_MESSAGE = 'No Hostname history present on settings';
+    export const CODEVERSION_INFO_MESSAGE = 'No Code Version history present on settings';
+    export const UPDATE_FILE_ERROR_MESSAGE = 'Error when updating dw.json file: ';
+    export const CREATE_FILE_ERROR_MESSAGE = 'Error when creating dw.json file: ';
+    export const CREATE_FILE_SUCCESS_MESSAGE = 'Created a dw.json file on this project folder';
+    export const FIX_FILE_ERROR_MESSAGE = 'Error on fixing the dw.json file: ';
+    export const FIX_FILE_SUCCESS_MESSAGE = 'The dw.json file was been fixed';
 }