+++ CHANGELOG.md
@@ -1,9 +1,31 @@
 # Change Log
 
-All notable changes to the "sfcc-dw-helper" extension will be documented in this file.
+## [0.0.14]
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- Minor fixes on sidebar layout
 
-## [Unreleased]
+## [0.0.13]
 
-- Initial release
+- Minor fixes on sidebar collapse feature
+
+## [0.0.12]
+
+- Implementation of history for **Hostname** and **Code Version** fields
+- Selection of user inputted history by **QuickPick**
+- Implementation of collapse sections on sidebar 
+- Possibility to execute Prophet commands 
+
+
+
+## [0.0.11]
+
+- Initial production release
+- Possibility to execute compiler commands 
+- Possibility to execute Prophet commands 
+
+
+
+
+## [0.0.1] - [0.0.10] 
+
+- Beta releases